cleaning up mygroup methods, and fix date to datetime exception in schedule view

+++ controllers/mygroup.py
@@ -39,6 +39,7 @@ def schedule():
                                                                                 db.cohort_plan.cohort_id,
                                                                                 db.cohort_plan.id)
         dbfile.write('%s : %s\n' % (datetime.datetime.now(), 'before for plan'))
+
         for plan in allPlans:
             if plan.cohort_plan.status=='new' and plan.cohort_plan.start_date < datetime.datetime.now().date():
                 plan.cohort_plan.status='active'
@@ -70,7 +71,6 @@ def schedule():
         return dict(allPlans=allPlans, allProgress=allProgress, allUsers=allUsers,
                     allComments=allComments, cohortName=cohortName)
 
-
 def newschedule():
     if auth.user == None:
         redirect(URL('default', 'user/login'))
@@ -89,28 +89,36 @@ def modifiedschedule():
 
 
 def modify():
-    db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).update(status='new')
-    db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).update(start_date=request.vars.startDate)
-    db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).update(end_date=request.vars.endDate)
-    db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).update(note=request.vars.note)
-    db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).update(created_by=auth.user)
-    db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).update(created_on=datetime.datetime.now())
-    plan = db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).select(db.cohort_plan.ALL).first()
-    plan_fields = db.cohort_plan._filter_fields(plan)
-    plan_fields['plan_id'] = plan.id
-    db.cohort_plan_revisions.insert(**plan_fields)
-
+    if auth.user == None:
+        redirect(URL('default', 'user/login'))
+    else:
+        db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).update(status='new',
+            start_date=datetime.datetime.strptime(request.vars.startDate, "%Y-%m-%d %H:%M:%S"),
+            end_date=datetime.datetime.strptime(request.vars.endDate, "%Y-%m-%d %H:%M:%S"),
+            note=request.vars.note,
+            created_by=auth.user,
+            created_on=datetime.datetime.now())
+        plan = db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).select(db.cohort_plan.ALL).first()
+        plan_fields = db.cohort_plan._filter_fields(plan)
+        plan_fields['plan_id'] = plan.id
+        db.cohort_plan_revisions.insert(**plan_fields)
 
 def delete():
-    db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).update(status='notStarted')
-    plan = db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).select(db.cohort_plan.ALL).first()
-    db.cohort_plan_revisions.insert(**db.cohort_plan._filter_fields(plan))
-    db((db.cohort_plan_revisions.chapter_id==request.vars.chapter) & (db.cohort_plan_revisions.cohort_id==auth.user.cohort_id)).update(plan_id=plan.id)
-    db((db.user_comments.cohort_id==auth.user.cohort_id) & (db.user_comments.chapter_id==request.vars.chapter)).delete()
+    if auth.user == None:
+        redirect(URL('default', 'user/login'))
+    else:
+        db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).update(status='notStarted')
+        plan = db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).select(db.cohort_plan.ALL).first()
+        db.cohort_plan_revisions.insert(**db.cohort_plan._filter_fields(plan))
+        db((db.cohort_plan_revisions.chapter_id==request.vars.chapter) & (db.cohort_plan_revisions.cohort_id==auth.user.cohort_id)).update(plan_id=plan.id)
+        db((db.user_comments.cohort_id==auth.user.cohort_id) & (db.user_comments.chapter_id==request.vars.chapter)).delete()
 
 def complete():
-    db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).update(status='completed', actual_end_date=datetime.datetime.now())
-    db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).update(actual_end_date=datetime.datetime.now())
+    if auth.user == None:
+        redirect(URL('default', 'user/login'))
+    else:
+        db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).update(status='completed', actual_end_date=datetime.datetime.now())
+        db((db.cohort_plan.chapter_id==request.vars.chapter) & (db.cohort_plan.cohort_id==auth.user.cohort_id)).update(actual_end_date=datetime.datetime.now())
 
 def comment():
 	db.user_comments.insert(cohort_id=auth.user.cohort_id,chapter_id=request.vars.chapter,comment=request.vars.text,comment_by=auth.user)
